airflow/requirements.txt
apache-airflow-providers-docker
PyYAML
apache-airflow-providers-postgres
psycopg2-binary
pyyaml
